Chasma-194: Updated status page to include execute shell command page
