--- chasma-thin-client/src/components/modals/ExecuteShellCommandsModal.tsx
@@ -1,221 +1,0 @@
-﻿import React, {useCallback, useState} from "react";
-import {Row} from "../types/CustomTypes"
-import {
-    ExecuteShellCommandRequest,
-    ShellCommandResult
-} from "../../API/ChasmaWebApiClient";
-import {isBlankOrUndefined} from "../../stringHelperUtil";
-import {shellClient} from "../../managers/ApiClientManager";
-
-/**
- * The members of the execute shell commands modal.
- */
-interface IExecuteShellCommandsProps {
-    /** Function to call when the modal is being closed. **/
-    onClose: () => void,
-
-    /** The repository identifier to execute shell commands in. **/
-    repositoryId: string | undefined,
-
-    /** Function to call when the response is successful. **/
-    onSuccess: () => void,
-}
-
-/**
- * Initializes a new ExecuteShellCommandsModal class.
- * @param props The properties to execute shell commands in repositories.
- * @constructor
- */
-const ExecuteShellCommandsModal: React.FC<IExecuteShellCommandsProps> = (props: IExecuteShellCommandsProps) => {
-    /** Gets or sets the shell command rows. **/
-    const [rows, setRows] = useState<Row[]>([]);
-
-    /** Gets or sets the error message. **/
-    const [errorMessage, setErrorMessage] = React.useState<string | undefined>(undefined);
-
-    /** Gets or sets a flag indicating whether the commands have been executed. **/
-    const [commandsExecuted, setCommandsExecuted] = React.useState<boolean>(false);
-
-    /** Gets or sets the command output after the commands have been executed. **/
-    const [output, setOutput] = useState<{
-        executedCommand: string | undefined;
-        success: boolean | undefined;
-        message?: string | undefined;
-    }[] | undefined>([]);
-
-    /** Adds a shell command row to the form. **/
-    const addCustomShellCommandRow = () => {
-        setRows(prev => [
-            ...prev,
-            {id: crypto.randomUUID(), first: "", second: ""}
-        ])};
-
-    /** Handles custom shell command row changes in the form. **/
-    const handleShellCommandChange = (
-            id: string,
-            field: "first" | "second",
-            value: string
-    ) => {
-            setRows(prev =>
-                prev.map(row =>
-                    row.id === id ? { ...row, [field]: value } : row
-                )
-            );
-        };
-
-   /**
-   * Handles the request execute custom commands in the API shell.
-   **/
-   const handleExecuteShellCommandsRequest = useCallback(async (e: React.FormEvent) => {
-       e.preventDefault();
-       const request = new ExecuteShellCommandRequest();
-       request.repositoryId = props.repositoryId;
-       request.commands = [];
-       rows.forEach(row => {
-       if (!isBlankOrUndefined(row.first) && request.commands) {
-           request.commands.push(row.first);
-        }
-       });
-
-       try {
-           const response = await shellClient.executeShellCommands(request);
-           if (response.isErrorResponse) {
-               setErrorMessage(response.errorMessage);
-               setCommandsExecuted(false);
-               setOutput([]);
-               return;
-           }
-
-           setOutput(
-               response.results?.map((result: ShellCommandResult) => ({
-                   executedCommand: result.executedCommand,
-                   success: result.isSuccess,
-                   message: result.outputMessage,
-               }))
-           );
-           props.onSuccess();
-       } catch (e) {
-           console.error(e);
-           setOutput([])
-           setErrorMessage("Error executing shell commands. Check console logs for more information.")
-       }
-       finally {
-           setCommandsExecuted(true);
-       }}, [rows]);
-
-    /** Resets the form to a pre-filled state. **/
-    function resetForm() {
-        setRows([]);
-        setOutput([]);
-        setCommandsExecuted(false);
-        setErrorMessage(undefined);
-    }
-
-    /**
-     * Deletes the row with the specified row identifier.
-     * @param rowId The row identifier.
-     */
-    function deleteShellCommandRow(rowId: string) {
-        const filteredRows = rows.filter(row => row.id !== rowId);
-        setRows(filteredRows);
-    }
-
-    return (
-            <>
-                <div className="modal-backdrop" onClick={props.onClose}>
-                    <div className="modal" onClick={(e) => e.stopPropagation()}>
-                        {!commandsExecuted && (
-                            <>
-                                <div className="command-row modern-input-row">
-                                    {errorMessage && <h3 className="modal-message">{errorMessage}</h3>}
-                                    <h2>Enter shell commands:</h2>
-                                    <button
-                                        type="button"
-                                        className="add-button modern-add"
-                                        onClick={addCustomShellCommandRow}
-                                    >
-                                        +
-                                    </button>
-                                </div>
-                                <br/>
-                                <div>
-                                    {rows.map(row => (
-                                        <div
-                                            key={row.id}
-                                            style={{
-                                                display: "flex",
-                                                gap: "10px",
-                                                marginBottom: "10px"
-                                            }}
-                                        >
-                                            <input
-                                                type="text"
-                                                placeholder="Example: git log --oneline"
-                                                className="command-input modern-input"
-                                                value={row.first}
-                                                onChange={e => handleShellCommandChange(row.id, "first", e.target.value)}/>
-                                            <button
-                                                className="remove-button modern-remove"
-                                                type="button"
-                                                onClick={() => deleteShellCommandRow(row.id)}
-                                            >
-                                                -
-                                            </button>
-                                        </div>
-                                    ))}
-                                    <br/>
-                                </div>
-                                <div className="modal-actions">
-                                    <button className="modal-button primary"
-                                            onClick={handleExecuteShellCommandsRequest}
-                                    >
-                                        Execute Commands
-                                    </button>
-                                    <button className="modal-button secondary"
-                                            onClick={props.onClose}
-                                    >
-                                        Close
-                                    </button>
-                                </div>
-                            </>
-                        )}
-                        {commandsExecuted && (
-                            <>
-                                <div className="modal-title">
-                                    <h2>Console Output</h2>
-                                </div>
-                                <br/>
-                                <div className="output-window">
-                                    {output?.map((result, index) => (
-                                        <div
-                                            key={index}
-                                            className={`output-entry ${result.success ? "success" : "failure"}`}
-                                        >
-                                            <span className="output-command">&gt; {result.executedCommand}</span>
-                                            {result.message && (
-                                                <div className="output-stdout">{result.message}</div>
-                                            )}
-                                        </div>
-                                    ))}
-                                </div>
-                                <div className="modal-actions">
-                                    <button className="modal-button primary"
-                                            onClick={resetForm}
-                                    >
-                                        Restart
-                                    </button>
-                                    <button className="modal-button secondary"
-                                            onClick={props.onClose}
-                                    >
-                                        Close
-                                    </button>
-                                </div>
-                            </>
-                        )}
-                    </div>
-                </div>
-            </>
-        )
-}
-
-export default ExecuteShellCommandsModal;
