+++ chasma-thin-client/src/styles/pages/batchOperationsPage.css
@@ -179,6 +179,10 @@
     border-radius: 16px;
     box-shadow: 0 8px 20px rgba(0,0,0,0.35);
     font-family: "Fira Code", "Source Code Pro", monospace;
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    min-height: 0;
 }
 
 .output-header {
@@ -202,22 +206,6 @@
 .clear-output-button:hover {
     transform: translateY(-1px);
     background: #616161;
-}
-
-.output-window {
-    max-height: 400px;
-    overflow-y: auto;
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-    background: #111827;
-    padding: 12px;
-    border-radius: 12px;
-    font-size: 0.95rem;
-    line-height: 1.5;
-    color: #e0e0e0;
-    scrollbar-width: thin;
-    scrollbar-color: #4f4b4b #1f2937;
 }
 
 .output-window::-webkit-scrollbar {
@@ -285,7 +273,9 @@
 }
 
 .output-window {
-    max-height: 400px;
+    min-height: 120px;
+    max-height: 60vh;
+    overflow-y: auto;
     overflow-y: auto;
     display: flex;
     flex-direction: column;
@@ -299,6 +289,9 @@
     color: #e0e0e0;
     scrollbar-width: thin;
     scrollbar-color: #4f4b4b #1f2937;
+    flex: 1;
+    overflow-y: auto;
+    min-height: 0;
 }
 
 .output-window::-webkit-scrollbar {
@@ -323,10 +316,15 @@
     font-family: "Fira Code", monospace;
     font-size: 0.95rem;
     white-space: pre-wrap;
-    word-break: break-word;
+    word-break: normal;
+    overflow-wrap: anywhere;
     position: relative;
     overflow-x: auto;
     transition: background 0.15s ease;
+    line-height: 1.6;
+    width: 100%;
+    display: block;
+    flex-shrink: 0;
 }
 
 .output-entry.success::before,
